se mueven funciones a nuevo complemento de resultados
